Fix version capture for developer installs
