+++ environment.yaml
@@ -11,6 +11,7 @@ dependencies:
   - python-casacore
   - pyyaml
   - scipy
+  - setuptools_scm
   - sphinx
   - coverage
   - pytest-cov
